pyproject.toml
[tool.poetry]
name = "v2i-pc-vis"
version = "0.1.0"
description = "Visualizer for fusing point clouds from vehicles with infrastructure."
authors = ["Ben Giacalone <bensgiacalone@gmail.com>"]
readme = "README.md"
packages = [{include = "v2i_pc_vis"}]

[tool.poetry.dependencies]
python = "^3.9,<3.13"
rerun-sdk = "^0.14.0"
cython = "^3.0.8"
open3d = "^0.18.0"
pip = "^24.0"
install = "^1.3.5"
open3d-cpu = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
